mac-u2s2@2020-03-14_04-17-35:

+++ detectron/utils/boxes.py
@@ -66,7 +66,7 @@ def boxes_area(boxes):
 def unique_boxes(boxes, scale=1.0):
     """Return indices of unique boxes."""
     v = np.array([1, 1e3, 1e6, 1e9])
-    hashes = np.round(boxes * scale).dot(v)
+    hashes = np.round(boxes * scale).dot(v).astype(np.int)
     _, index = np.unique(hashes, return_index=True)
     return np.sort(index)
